requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.0
pandas==2.1.1
pyodbc==5.0.1
python-dateutil==2.8.2
mysql-connector-python==8.0.5
flask_jwt_extended
bcrypt
PyYAML==6.0.1
